separate util.c into util_system.c and util.c
