+++ src/cmd-run.c
@@ -219,7 +219,7 @@ int cmd_run_star(int argc,char **argv,struct sub_command* cmd) {
         cmd=s_cat(cmd,q(" "),q("\""),escape_string(arg[i]),q("\""),NULL);
       }
       SetConsoleCtrlHandler(ConsoleCtrlHandler, TRUE);
-      exit(system(cmd));
+      exit(System(cmd));
       s(cmd);
     }
 #else
