+++ src/util.h
@@ -133,6 +133,7 @@ int delete_file(char* pathspec);
 int rename_file(char* file,char* new_name);
 void touch(char* path);
 char* system_(char* cmd);
+int System(const char* command);
 char** parse_cmdline(char* cmdline,int *argc);
 int free_cmdline(char** argv);
 int system_redirect(const char* cmd,char* filename);
